requirements.txt
pypdf
tqdm
sentence-transformers
faiss-cpu
